refactor: 에러 메시지 상수를 따로 분리
